refactor : 멤버 도메인에 포인트가 있던 기능 리팩토링. 일단 돌아가긴 한다...후

+++ apps/commerce-api/src/main/java/com/loopers/application/member/command/PointChargeCommand.java
@@ -1,7 +1,9 @@
 package com.loopers.application.member.command;
 
+import java.math.BigDecimal;
+
 public record PointChargeCommand(
         Long memberId,
-        Long amount
+        BigDecimal amount
 ) {
 }
